move SchedulerHelper to a separate file
